+++ resources/js/api.js
@@ -391,16 +391,23 @@ export async function editComment(cid, content, endpoint = '/comment/{cid}') {
     );
 };
 
-export async function addOrUpdateBibliographyItem(item) {
-    let data = {};
+export async function addOrUpdateBibliographyItem(item, file) {
+    const data = new FormData();
     for(let k in item.fields) {
-        data[k] = item.fields[k];
+        data.append(k, item.fields[k]);
     }
-    data.type = item.type.name;
+    data.append('type', item.type.name);
+    if(file) {
+        if(file == 'delete') {
+            data.append('delete_file', true);
+        } else {
+            data.append('file', file);
+        }
+    }
 
     if(item.id) {
         return $httpQueue.add(
-            () => http.patch(`bibliography/${item.id}`, data).then(response => response.data)
+            () => http.post(`bibliography/${item.id}`, data).then(response => response.data)
         );
     } else {
         return $httpQueue.add(
@@ -779,6 +786,12 @@ export async function removeEntityTypeAttribute(etid, aid) {
 export async function deleteAttribute(aid) {
     return $httpQueue.add(
         () => http.delete(`editor/dm/attribute/${aid}`).then(response => response.data)
+    );
+};
+
+export async function deleteBibliographyItemFile(id) {
+    return await $httpQueue.add(
+        () => http.delete(`bibliography/${id}/file`).then(response => response.data)
     );
 };