+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 - Info modal in references tab to view all information from an literature reference
 - Copy BibTeX-Code of literature item to clipboard
 - Entity type color dot to entity detail's link list (entity type name is shown on hover)
+- Attach files to bibliography entries
+- Validation to bibliography add/edit form
+- More (optional) bibtex fields (doi, email, url, type)
 ### Fixed
 - Only delete avatar if is set and exists on disk
 - Display image in About modal
@@ -13,6 +16,8 @@ All notable changes to this project will be documented in this file.
 - Display errors from server if user can not be added (e.g. due to already existing nickname or email address)
 - References list not visible after added a new one if it's the first literature reference added to that attribute
 - Newly created entity types only selectable as sub-entity types after reload
+- Column ordering in bibliography
+- Saving changed type of bibtex entry
 ### Changed
 - Updated Modal Library to Version 4
 - Better handling of `"`, `{` and `{{` in BibTeX